Changed types from int to double, fixed sum method + upgraded readme

+++ testApp/SuitesConsole/class/Arithmetique.cs
@@ -1,11 +1,11 @@
 public class Arithmethique : Suite
 {
-    public Arithmethique(int raison, int premierTerme) : base(raison, premierTerme) {}
+    public Arithmethique(double raison, double premierTerme) : base(raison, premierTerme) {}
 
 
-    public override int GetValueAt(int rank) => PremierTerme + (rank * Raison);
+    public override double GetValueAt(int rank) => PremierTerme + (rank * Raison);
 
-    public override long GetSum(int firstN) // Peut on return un decimal ?
+    public override double GetSum(int firstN) // Peut on return un decimal ?
     {
         return firstN * ((PremierTerme + GetValueAt(firstN-1)) / 2);
     }
